Web Inspector: Uncaught Exception: TypeError: snapshot.element.toDataURL is not a function. (In 'snapshot.element.toDataURL()', 'snapshot.element.toDataURL' is undefined) https://bugs.webkit.org/show_bug.cgi?id=302234
Reviewed by BJ Burg.

`OffscreenCanvas` doesn't have a `toDataURL` method (nor does it have a `toBlob`, though there is a close-but-not-identical `convertToBlob`).

Additionally, in order to preview the `OffscreenCanvas` we need to either transfer to an `ImageBitmap` (which means we can't reuse the state of the `OffscreenCanvas` later on) or have it be created from a `<canvas>` (which already works).

* Source/WebInspectorUI/UserInterface/Models/Recording.js:
(WI.Recording.prototype.createContext.createCanvasContext):
(WI.Recording.prototype.createContext.createOffscreenCanvasContext):
(WI.Recording.prototype.async _process):
* Source/WebInspectorUI/UserInterface/Views/RecordingContentView.js:
(WI.RecordingContentView.prototype._generateContentCanvas2D):

Canonical link: https://commits.webkit.org/303229@main