REGRESSION(302840@main): [GStreamer] Broke webcodecs tests https://bugs.webkit.org/show_bug.cgi?id=302590
Reviewed by Xabier Rodriguez-Calvar.

The issues were mainly due to caps negotiation errors on the encoder. After setting a framerate on the
encoder, the input sample caps should also be coherent with the new framerate.

Driving-by, disallow vavp8alphadecodebin for VP8 decoding, it is not suitable for WebCodecs.

Canonical link: https://commits.webkit.org/303245@main