Mark AuxiliaryProcessProxy::fromConnection as nullable https://bugs.webkit.org/show_bug.cgi?id=302643 rdar://164887710
Reviewed by Geoffrey Garen.

This marks this API as returning a value which may potentially be null. This is
necessary for Swift/C++ interoperability to work properly; for our current
prototypes of a CoreIPC message receiver in Swift it's the only such API which
needs to be so annotated.

This suppresses a related warning in cmake builds which is already suppressed
in Xcode builds.

Canonical link: https://commits.webkit.org/303248@main