[blocks-in-inline] Fix local previous/nextLinePosition to adjust for local offset https://bugs.webkit.org/show_bug.cgi?id=302777 rdar://165041221
Reviewed by Alan Baradlay.

Blocks-in-inline revealed an existing bug in these functions.
If the selection was extended to an inline block we would select wrong line inside it because
bad coordinates. This becomes more important with blocks-in-inline.

Test: editing/deleting/pruning-after-merge-3.html
* LayoutTests/editing/deleting/pruning-after-merge-3-expected.html: Added.
* LayoutTests/editing/deleting/pruning-after-merge-3.html: Added.

Add a test where we extend selection to an inline-block.
The tested behavior matches Chrome.

* Source/WebCore/editing/VisibleUnits.cpp:
(WebCore::previousLinePosition):
(WebCore::nextLinePosition):

visiblePositionForPoint() expects point in the local coordinates of the renderer.

Canonical link: https://commits.webkit.org/303253@main