Fix linked-on-or-after check for using process assertions off the main thread https://bugs.webkit.org/show_bug.cgi?id=302758 rdar://164866073
Reviewed by Chris Dumez.

In 302397@main, I passed in the wrong SDKAlignedBehavior for the check on whether to fault or assert
when manipulating process assertions off the main thread.

This fixes that issue, and also renames the symbol to make it clearer that the crash is due to an
error in the application rather than an issue in WebKit. The existing assert with message macro
doesn't seem to always result in a message in the crash log from what I've seen.

* Source/WebKit/UIProcess/ProcessThrottler.cpp:
(WebKit::crashDueToApplicationCallingMainThreadOnlyWebKitAPIFromBackgroundThread):
(WebKit::assertIfCalledFromBackgroundThread):

Canonical link: https://commits.webkit.org/303266@main