StabilityTracer: WebAutomationSession::deleteAllCookies must not send null strings over IPC https://bugs.webkit.org/show_bug.cgi?id=302488 rdar://164526691
Reviewed by BJ Burg.

It seems that in NetworkStorageSession::deleteCookiesForHostnames, one of the Strings
in hostnames is a null String, so we crash when trying to add it to hostnamesSet.

From the crash trace, hostnames is coming from the IPC sent to
WebCookieManager::deleteCookiesForHostnames which is sent from
HTTPCookieStore::deleteCookiesForHostnames, which is called from
WebAutomationSession::deleteAllCookies.

In WebAutomationSession::deleteAllCookies, if host is a null String (and it seems
possible that the URL returned by PageLoadState::activeURL could have a null host),
then hostnames will contain null Strings.

To fix this, we ensure that if host is a null String, the IPC is not sent.

* Source/WebKit/UIProcess/Automation/WebAutomationSession.cpp:
(WebKit::WebAutomationSession::deleteAllCookies):

Canonical link: https://commits.webkit.org/303282@main