Restore SOAuthorizationOptionInitiatorOrigin to pre-rdar://144600565 behavior rdar://154104792
Reviewed by David Kilzer.

In SOAuthorizationSession::continueStartAfterDecidePolicy we have code that used the nullness of
API::NavigationAction::sourceFrame to determine whether to have an empty string as the value of
SOAuthorizationOptionInitiatorOrigin in the dictionary.  rdar://144600565 changed the nullness,
so in order to get the same behavior we also need to check whether the security origin is opaque.
Otherwise we get "null" as the origin from SecurityOrigin::toString which the user of
SOAuthorizationOptionInitiatorOrigin interprets differently.

* Source/WebKit/UIProcess/Cocoa/SOAuthorization/SOAuthorizationSession.mm:
(WebKit::SOAuthorizationSession::continueStartAfterDecidePolicy):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SOAuthorizationTests.mm:
(TestWebKitAPI::TEST(SOAuthorizationRedirect, AuthorizationOptionsAboutBlank)):

Originally-landed-as: 297297.300@safari-7622-branch (0ee9dc82a89f). rdar://164280467
Canonical link: https://commits.webkit.org/303305@main