Do not expose WKScrollGeometryAdapter.swift initialiser in C++ https://bugs.webkit.org/show_bug.cgi?id=302642 rdar://164465358
Reviewed by Richard Robinson.

This excludes the initialiser of WKScrollGeometryAdapter.swift from Swift/C++
interoperability, which is necessary to prevent compile errors because it
depends on objective-C types. Generally this sort of manual exclusion isn't
necessary, but in this case this works around a bug in the header generation.

Canonical link: https://commits.webkit.org/303320@main