[JSC] Fold `new Uint8Array(new ArrayBuffer(length)).length` to `length` https://bugs.webkit.org/show_bug.cgi?id=302832 rdar://165095809
Reviewed by Justin Michaud.

We found that this is particularly common since Uint8Array is byte view
to the underlying ArrayBuffer. This patch introduces the above folding
so that we can make GetUndetachedTypeArrayLength non-opaque to the loop
when it is using `length` directly for the loop condition instead of
`view.length`. We also make it more precise about MayExit for PutByVal
with TypedArray in FTL. Since bound-check is already separate, if it is
typed array, then it does not exit within the node.

* JSTests/stress/bound-check-removal-typed-array-byte-length.js: Added.
(shouldBe):
(test):
* Source/JavaScriptCore/dfg/DFGMayExit.cpp:
* Source/JavaScriptCore/dfg/DFGStrengthReductionPhase.cpp:
(JSC::DFG::StrengthReductionPhase::handleNode):

Canonical link: https://commits.webkit.org/303351@main