Array allocation sinking materialize needs to work without exitOK https://bugs.webkit.org/show_bug.cgi?id=302747 rdar://164644155
Reviewed by Justin Michaud and Yusuke Suzuki.

Right now we can fail validation when ObjectAllocationSinking ends up
materializing an Array at an exit invalid node. This is because the
PutByVal we used reports that it could exit in mayExit. This is really
a false positive because we can never store out of bounds and the edges
should already be checked.

This patch makes PutByValAlias report that it does not exit (unless
an edge speculation would fail or for TypeArrays) no longer report that
it exits to work around this. It is already the case that PutByValAlias
does not exit (except for TypedArrays although that could be fixed).

To clarify this new use with ObjectAllocationSinking PutByValAlias
has also been renamed to PutByValDirectResolved.

Test: JSTests/stress/array-sink-materialize-cycle-break-in-exit-invalid.js
Canonical link: https://commits.webkit.org/303353@main