In IPInt calls, save the pointer to the first non-arg stack slot as a frame-relative value https://bugs.webkit.org/show_bug.cgi?id=302877 rdar://165141831
Reviewed by Dan Hecht.

As part of the call sequence, IPInt computes the address of the first non-argument Wasm
stack slot and stores it on the stack. After the call the address is used while pushing
the return values. To simplify JSPI implementation, we should store the address as a value
relative to the frame pointer, so stack data can be easily moved to a different address.

The patch makes this change together with comment updates to keep them current.

Tests: covered by existing ones.
Canonical link: https://commits.webkit.org/303359@main