[GLIB] Move the run loop observer repeating logic to ActivityObserver https://bugs.webkit.org/show_bug.cgi?id=302855
Reviewed by Nikolas Zimmermann.

This way neither the caller to notify() nor the callback need to check
whether the observer is repeating or not. It's simpler, but also safer
because ActivityObserver is refcounted and always protected while
notify() is called, so we can safely check if it's repeating after the
callback is called even if it's stopped by the callback.

Canonical link: https://commits.webkit.org/303382@main