Drag image contains content from neighboring elements when css transform is involved
https://bugs.webkit.org/show_bug.cgi?id=302344

Reviewed by Simon Fraser.

Skips painting content of elements that are not descendants of the dragged
element.  Specifically, sets shouldPaintContent to false when isLayerInSubtree
is false.

This builds on related fixes in a3e5281 and 1d76f58 from a few months ago.
I think the latter commit unintentionally mixed up the logic when it reworked
the former.

Test: fast/snapshot/scaled-down-drag.html

* LayoutTests/fast/snapshot/scaled-down-drag-expected.html: Added.
* LayoutTests/fast/snapshot/scaled-down-drag.html: Added.
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintLayerContents):

Canonical link: https://commits.webkit.org/303407@main