[threaded-animations] https://scroll-driven-animations.style/demos/scroll-shadows/css/ asserts with "Threaded Scroll-driven Animations" enabled https://bugs.webkit.org/show_bug.cgi?id=302953
Reviewed by Alan Baradlay.

We made a mistake in 303312@main and removed the requirement that an effect's
associated animation not be pending to qualify for acceleration and as a result
we hit the `ASSERT(!animation->pending())` assertion in the `AcceleratedEffect`
constructor.

We simply reinstate this requirement in `KeyframeEffect::canBeAccelerated()`
for threaded animations.

Test: webanimations/threaded-animations/pending-animation-crash.html

* LayoutTests/webanimations/threaded-animations/pending-animation-crash-expected.txt: Added.
* LayoutTests/webanimations/threaded-animations/pending-animation-crash.html: Added.
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::canBeAccelerated const):

Canonical link: https://commits.webkit.org/303410@main