[libpas] Extend Retag-on-Scavenge to cover bitfit objects as well rdar://156723433 https://bugs.webkit.org/show_bug.cgi?id=300151
Reviewed by Dan Hecht.

This patch finishes the retag-on-scavenge work by extending the
implementation to cover bitfit objects as well.
The main issue with the prior implementation is that it would attempt
to retag even for calls to pas_bitfit_page_deallocate_with_page_impl
that were not actually *deallocating*, but simply shrinking the object
or checking its size. This patch fixes that by guarding the retag
behind a switch as appropriate.

Canonical link: https://commits.webkit.org/303411@main