AudioVideoRenderer will constantly request for more data even if earlier request not serviced. https://bugs.webkit.org/show_bug.cgi?id=303001 rdar://165259808
Reviewed by Eric Carlson.

The AVSampleBufferAudioRenderer can call the callback provided to `requestMediaDataWhenReadyOnQueue`
multiple times before stopRequestingMediaData is called.
When used across processes, it may call spuriously the callback before the
message to stop the request comes back.
There should be no need for the AudioVideoRenderer to ask to stop the requestMediaDataWhenReadyOnQueue
after each callback. So we make it part of the API.
The callback will now be called once, until the next requestMediaDataWhenReadyOnQueue
is issued.

Fly-By: Typo (m_readyToRequestVideoData vs m_readyToRequestAudioData) could
have caused audio callback to not be called if we were waiting on a video
flush.

* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.mm:
(WebCore::AudioVideoRendererAVFObjC::requestMediaDataWhenReady):
* Source/WebCore/platform/graphics/avfoundation/objc/SourceBufferPrivateAVFObjC.mm:
(WebCore::SourceBufferPrivateAVFObjC::didBecomeReadyForMoreSamples):
* Source/WebCore/platform/graphics/cocoa/MediaPlayerPrivateWebM.mm:
(WebCore::MediaPlayerPrivateWebM::didBecomeReadyForMoreSamples):

Canonical link: https://commits.webkit.org/303456@main