[WTF] WorkerPool::Worker should use a CheckedPtr instead of a CheckedRef for the WorkerPool reference https://bugs.webkit.org/show_bug.cgi?id=302929
Reviewed by Nikolas Zimmermann.

Since AtomicThread keeps a reference of itself while the thread is
running, it can happen that the WorkerPool is destroyed before the
thread actually finishes even if join was called and threadIsStopping is
already called. This can happen easily when the system is under heavy
load, as we have seen in the bots when running the layout tests.

Canonical link: https://commits.webkit.org/303468@main