Add Swift wrapper header https://bugs.webkit.org/show_bug.cgi?id=302931 rdar://165200310
Reviewed by Geoffrey Garen and Elliott Williams.

In future various parts of WebKit's C++ will want to include the generated
definitions/declarations of Swift types and functions. This is already the case
with the Swift demo feature and will become more important with upcoming
CoreIPC work. To do this, C++ needs to refer to a header file emitted by Swift.
That header file has a few quirks which make it awkward to include:

* It generates warnings which WebKit treats as errors.
* It does not include pre-requisite headers, e.g. for the types used as
  parameters to Swift functions.
* The WebKit build process may generate the header with two different names
  depending on SDK version.

This commit introduces a 'wrapper' header which resolves all these problems in
one place, so that any other code wishing to refer to Swift definitions and
declarations can simply #include "WebKit-Swift.h". At the moment this is not a
dramatic simplification, but it will become so as Swift usage expands.

Canonical link: https://commits.webkit.org/303477@main