Remove excessive documentation of YarrSyntaxCheckable concept and update the concept https://bugs.webkit.org/show_bug.cgi?id=302968
Reviewed by Yusuke Suzuki.

This comment is no longer necessary since 290702@main, when it was
originally removed at the same time the YarrSyntaxCheckable concept was
added. The concept enforces at compile time that the type must be as
expected.

The comment was reintroduced in 293609@main but I'm not sure why. The
comment already has the wrong parameter types in atomCharacterClassAtom
and atomCharacterClassRange, since it's not possible to enforce that the
comment is updated when the concept changes. It's probably better to
just refer to the concept instead.

Canonical link: https://commits.webkit.org/303499@main