Double free / use-after-free of static FcPattern in FontPlatformDataFreeType::create https://bugs.webkit.org/show_bug.cgi?id=302858
Reviewed by Carlos Garcia Campos.

Here we adopt a reference that we do not own. Well, we sort of own it,
in the static local variable, but we need to *continue* owning it and
therefore must not pass ownership to the FontPlatformData constructor.

So, remove the adoptRef(). A RefPtr will be implicitly created,
increasing the ref count on the FcPattern, as is required.

(Note this FcPattern will be leaked, but that's OK because it's global
data. The ref count should be 1 at program termination.)

Canonical link: https://commits.webkit.org/303513@main