[WebRTC] traffic class should be set to VI if service class is set to VI rdar://165386600 https://bugs.webkit.org/show_bug.cgi?id=303083
Reviewed by Jean-Yves Avenard.

When we set service class to video interactive, we also need to set traffic to video interactive,
so as to benefit from network slicing.
For now, we force traffic class to video interactive when service class is enabled.
In the future, we might finetune this heuristic based on https://github.com/w3c/webrtc-extensions/pull/247.

Canonical link: https://commits.webkit.org/303528@main