[GStreamer] Using the `loop` property in a media element breaks the seeking and looping. https://bugs.webkit.org/show_bug.cgi?id=302541
Reviewed by Xabier Rodriguez-Calvar.

Seamless (segment) seeking is now disabled for audio players and media files muxed in Ogg
containers, due to missing support for that in oggdemux. Segment seeking is mostly relevant for
videos, in order to avoid a frame flashing before looping back to the beginning of the video.

This patch also improves seek support in looping videos, by making those seek events flush the
pipeline. Once such flush seek is done, another segment seek is performed, so that the seamless
rendering keeps working at the next loop iteration.

With this patch media/video-loop.html now runs in ~3.9 seconds, while without it was taking 10
seconds to run (because the seek near the end of video was not working as expected).

Canonical link: https://commits.webkit.org/303539@main