StabilityTracer: com.apple.WebKit.GPU.Development at com.apple.WebKit:  IPC::MessageReceiver::~MessageReceiver https://bugs.webkit.org/show_bug.cgi?id=303040 rdar://148660972
Reviewed by Youenn Fablet.

If the web content process got terminated before the RemoteSourceBufferProxy's
destruction, we would get an assertion as `disconnect()` would have been
a no-op and it didn't remove itself from the connection's messageReceiverMap.

We now invalidate all RemoteMediaSource and their RemoteSourceBufferProxy when the connection
to the content process is terminated.

Canonical link: https://commits.webkit.org/303561@main