[GTK][WPE] Do not block while waiting for tiles in the threaded compositor https://bugs.webkit.org/show_bug.cgi?id=303087
Reviewed by Nikolas Zimmermann and Miguel Gomez.

Ensure we don't schedule a composition until all pending tiles are done
instead of synchronously waiting for tiles during the composition.
CoordinatedSceneState now keeps track of the ongoing tile painting
operations. A new state WaitingForTiles is added to the compositor so
that if there are pending tiles when a rendering update composition is
requested we switch to the WaitingForTiles state. The compositor is
notified by LayerTreeHost every time the pending tiles change, so if we
were waiting for tiles and there are no more pending tiles we switch to
Schedule state and continue normally. This is the previous behavior, but
asynchronously waiting for the tiles instead. This will allow to process
other composition requests while waiting for tiles of a rendering
update request.

Canonical link: https://commits.webkit.org/303564@main